-FIX, 3482207: copy all result column data to clipboard -CHG, 3098471: support v4.0 of .Net when only framework installed -CHG, 3424154/1816655: allow multiple search paths to be entered, separate full paths by use of , (comma) -CHG, 2078252: add ability to copy name, located in, and located in + name to clipboard -CHG, 1955653: set focus to the search text field on startup -CHG, 1947760: update the default exclude list to exclude images (.gif, .bmp, .png, .jpeg, .jgp) -CHG, 1561584: add ability to skip hidden/system files/directories -CHG, 3054576: add patch to exclude surrounding full file path with quotes when quotes are provided in the command line parameters for a text editor
-ADD, language support for context menu on file list/size column/min,max size/modified date before,after
-ADD, restore default extensions button to restore file exclude extensions list to default list
-ADD, output all search options to html, xml.  Update txt to include brief summary header

+++ AstroGrep/libAstroGrep/IFileFilterSpec.cs
@@ -29,11 +29,11 @@ namespace libAstroGrep
       /// <summary>
       /// Minimum file size
       /// </summary>
-      int FileSizeMin { get; }
+      long FileSizeMin { get; }
 
       /// <summary>
       /// Maximum file size
       /// </summary>
-      int FileSizeMax { get; }
+      long FileSizeMax { get; }
    }
 }
